:bug: Fixes enum import bug
