+++ www/src/types.ts
@@ -1,0 +1,9 @@
+export enum CorePlugins {
+  Anchor = 1 << 0,
+  Collapse = 1 << 1,
+  Focus = 1 << 2,
+  Intersect = 1 << 3,
+  Mask = 1 << 4,
+  Morph = 1 << 5,
+  Persist = 1 << 6,
+}
